DeathMSG+Respawn+new Secutiry camera
Implemented a "lose game" message. When the character enters a situation
where it loses, the game shows a message to it.
Also, when it loses, the game respawns a few seconds later.
And removed the red cones from the security cameras, switching it for a
light. Now you can actually see what is in front of you, not a giant
cone of death.
Still need to change other cameras and include
buttons/portals/obstacles.