PROrtals implemented!
Portals with renderTexture implemented!
How to use:
Get a GameObject from the prefab "PROrtalPREFAB"
Rename this game object (so we can understand which portal it is)
Create a RenderTexture and a Material for each portal (with names that
help identifying them)
Assign the correct texture to the correct camera AND to the correct
material; them the correct material to its portal!
UnityPRO s2