Button and Lights OK!
Created a new script to deactivate some GameObjects on the sceneLoad.
This way we can toggle  cameras/lights/whatever we want between ACTIVE
and DEACTIVE. No direction toggles yet.
Created a PREFAB for the security Cameras. If they start deactive, just
add the StartDeactivated script to them. If they toggle, add the tag
"Light".
Created a PREFAB for the buttons and modified it a bit. Now you just
need to press "e" near the button and it will do it work.
Almost finished the prototype. Need only to work on MirrorWorld.